[CMake] Fixes issues with DUNE_INSTALL_MODULEDIR for changes in INSTALL variables
Previously, if the user changed e.g. CMAKE_INSTALL_PREFIX or
CMAKE_INSTALL_DATAROOTDIR these changes where never mimiced in
DUNE_INSTALL_MODULEDIR. This patch fixes this problem, such that
DUNE_INSTALL_MODULEDIR changes unless it is explicitely set by the
user before.

We do using the same approach as in GNUInstallDirs.cmake. That is, we
whether DUNE_INSTALL_MODULEDIR is set to a non-empty string. If that
is the case we set an empty cache variable and overwrite locally with
the default value. The default value is now a relative path.