Move stack-compatible shared_from_this to a separate class and header
The stack-compatible shared_from_this is useful beyond the DifferentiableFunction
and it also clutters up its interface with implementation details, so this patch
moves it to a separate class and header.

stack_compatible_shared_from_this works just like std::enable_shared_from this:
Just inherit from it to gain the functionality.

In the long run, the class should move to dune-common, but that requires writing
additional tests and will have to wait.