[DenseMatrix] Introduce DenseMatrixAssigner for initializing dense matrices from arbitrary right hand sides
In order to overcome the deficiencies of istl_assign_to_fmatrix() we
propose to always use a struct DenseMatrixAssigner for initializing a
dense matrix. This class can be specialized by users for any right hand
side to be used in an assignment.  The new mechanism can bei default
implemented using existing implementations of istl_assign_to_fmatrix()
as well.